\name{BosonSampling-package}
\alias{BosonSampling-package}
\docType{package}
\title{
\packageTitle{BosonSampling}
}
\description{
\packageDescription{BosonSampling}
}
\details{
\packageIndices{BosonSampling}
}
\author{
Peter Clifford <peter.clifford@jesus.ox.ac.uk> and Raphaël Clifford <clifford@cs.bris.ac.uk>

Maintainer: \packageMaintainer{BosonSampling}
}
