\name{BosonSampling-package}
\alias{BosonSampling-package}
\docType{package}
\title{
\packageTitle{BosonSampling}
}
\description{
\packageDescription{BosonSampling}
}
\details{

\packageIndices{BosonSampling}
}
\author{
\packageAuthor{BosonSampling}

Maintainer:  \packageMaintainer{BosonSampling}
}
